+++ lib/features/azkar/azkar_tap/data/repos/azkar_tap_repo.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:islamic_app/core/errors/failure.dart';
 
-abstract class TasbeehTapRepo{
+abstract class AzkarTapRepo{
   Future<Either<Failure,List<String>>> fetchAzkarTitles();
 }
