+++ lib/features/azkar/azkar_tap/presentation/view/widget/display_list_azkar_title.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:islamic_app/core/utils/widgets/custom_elevated_button.dart';
-import 'package:islamic_app/features/tasbeeh_tap/presentation/view_model/tasbeeh_tap_view_model.dart';
+import 'package:islamic_app/features/azkar/azkar_tap/presentation/view_model/azkar_tap_view_model.dart';
 import 'package:provider/provider.dart';
-import '../../../../../core/utils/widgets/custom_circular_progress_indicator.dart';
+
+import '../../../../../../core/utils/widgets/custom_circular_progress_indicator.dart';
 
 class DisplayListAzkarTitle extends StatefulWidget {
   const DisplayListAzkarTitle({super.key});
@@ -15,11 +16,11 @@ class _DisplayListAzkarTitleState extends State<DisplayListAzkarTitle> {
   @override
   void initState() {
     super.initState();
-    Provider.of<TasbeehTapViewModel>(context,listen: false).fetchAzkarTitle();
+    Provider.of<AzkarTapViewModel>(context,listen: false).fetchAzkarTitle();
   }
   @override
   Widget build(BuildContext context) {
-    TasbeehTapViewModel tasbeehTapViewModel = Provider.of<TasbeehTapViewModel>(context);
+    AzkarTapViewModel tasbeehTapViewModel = Provider.of<AzkarTapViewModel>(context);
     if(tasbeehTapViewModel.azkarTitles.isEmpty){
       return const SliverFillRemaining(child: CustomCircularProgressIndicator());
     }else{
