+++ lib/features/azkar/azkar_tap/presentation/view/widget/tasbeeh_tap_body.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:islamic_app/features/tasbeeh_tap/presentation/view/widget/show_image_sabhuh.dart';
+import 'package:islamic_app/features/azkar/azkar_tap/presentation/view/widget/show_image_sabhuh.dart';
 
 import 'display_list_azkar_title.dart';
