+++ lib/features/azkar/azkar_tap/presentation/view_model/azkar_tap_navigator.dart
@@ -1,5 +1,5 @@
 import 'package:islamic_app/core/base/base_navigator.dart';
 
-abstract class TasbeehTapNavigator implements BaseNavigator{
+abstract class AzkarTapNavigator implements BaseNavigator{
 
 }
