+++ lib/features/home_layout/presentation/view_model/home_layout_view_model.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:islamic_app/core/base/base_view_model.dart';
+import 'package:islamic_app/features/azkar/azkar_tap/presentation/view/tasbeeh_tap_view.dart';
 import '../../../quran/quran_tap/presentation/view/quran_tap/view/quran_tap.dart';
-import '../../../tasbeeh_tap/presentation/view/tasbeeh_tap_view.dart';
 import '../../models/bottom_navigation_item_model.dart';
 import 'home_layout_navigator.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -17,7 +17,7 @@ class HomeLayoutViewModel extends BaseViewModel<HomeLayoutNavigator>{
     Container(
       color: Colors.transparent,
     ),
-    const TasbeehTapView(),
+    const AzkarTapView(),
     Container(
       color: Colors.transparent,
     ),
